backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.119.1",
    "fastapi-users>=14.0.1",
    "fastapi-users-db-sqlalchemy>=7.0.0",
    "libmagic>=1.0",
    "loguru>=0.7.3",
    "pydantic-settings>=2.11.0",
    "python-magic>=0.4.27",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]
